+++ JsonMidiCreator/TalkieSPI/src/MultiPlayer.hpp
@@ -62,7 +62,7 @@ protected:
     }
     
 
-    long _bpm_10 = 1200;
+    uint16_t _bpm_10 = 1200;
 
     bool command_run(const uint8_t command_index, JsonObject json_message) override {
         (void)json_message; // Silence unused parameter warning
@@ -75,7 +75,7 @@ protected:
 
     
     bool command_set(const uint8_t command_index, JsonObject json_message) override {
-        long json_value = json_message["v"].as<long>();
+        uint32_t json_value = json_message["v"].as<uint32_t>();
         switch (command_index)
         {
         case 1:
@@ -90,7 +90,7 @@ protected:
     }
 
     
-    long command_get(const uint8_t command_index, JsonObject json_message) override {
+    uint32_t command_get(const uint8_t command_index, JsonObject json_message) override {
         (void)json_message; // Silence unused parameter warning
         switch (command_index)
         {
